+++ plugins/common-repo/src/main/groovy/com/apgsga/common/repo/extensions/MavenSettingsExtension.groovy
@@ -1,24 +1,9 @@
-/*
- * Copyright 2014 the original author or authors.
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
+package com.apgsga.common.repo.extensions
 
-package net.linguica.gradle.maven.settings
 
 import org.gradle.api.Project
 
-class MavenSettingsPluginExtension {
+class MavenSettingsExtension {
     private Project project
 
     /**
@@ -27,7 +12,7 @@ class MavenSettingsPluginExtension {
      */
     String userSettingsFileName = System.getProperty("user.home") + "/.m2/settings.xml"
 
-    public setUserSettingsFileName(String userSettingsFileName) {
+    void setUserSettingsFileName(String userSettingsFileName) {
         this.userSettingsFileName = userSettingsFileName;
         LoaderManager lm = new LoaderManager(project);
         lm.load();
@@ -44,11 +29,11 @@ class MavenSettingsPluginExtension {
      */
     boolean exportGradleProps = true
 
-    MavenSettingsPluginExtension(Project project) {
+    MavenSettingsExtension(Project project) {
         this.project = project
     }
 
-    public File getUserSettingsFile() {
+    File getUserSettingsFile() {
         return project.file(userSettingsFileName)
     }
 }
