+++ plugins/common-repo/src/main/resources/META-INF/gradle-plugins/com.apgsga.common.repo.properties
@@ -1,0 +1,1 @@
+implementation-class=com.apgsga.common.repo.plugin.ApgCommonRepoPlugin
