--- plugins/common-repo/src/main/resources/META-INF/gradle-plugins/net.linguica.maven-settings.properties
@@ -1,1 +1,0 @@
-implementation-class=net.linguica.gradle.maven.settings.MavenSettingsPlugin
