+++ plugins/version-manager/src/main/kotlin/com/apgsga/maven/dm/plugin/VersionResolutionPlugin.kt
@@ -3,7 +3,7 @@ package com.apgsga.maven.dm.plugin
 import com.apgsga.maven.dm.ext.VersionResolutionExtension
 import com.apgsga.packaging.plugins.ApgCommonPackagePlugin
 import com.apgsga.revision.manager.domain.RevisionManagerBuilder
-import net.linguica.gradle.maven.settings.MavenSettingsPlugin
+import com.apgsga.common.repo.plugin.ApgCommonRepoPlugin
 import org.gradle.api.Action
 import org.gradle.api.Plugin
 import org.gradle.api.Project
@@ -13,7 +13,7 @@ import org.gradle.api.publish.maven.plugins.MavenPublishPlugin
 
 open class VersionResolutionPlugin : Plugin<Project> {
     override fun apply(project: Project) {
-        project.plugins.apply(MavenSettingsPlugin::class.java)
+        project.plugins.apply(ApgCommonRepoPlugin::class.java)
         project.plugins.apply(MavenPublishPlugin::class.java)
         project.plugins.apply(ApgCommonPackagePlugin::class.java)
         val revisionManagerBuilder = RevisionManagerBuilder.create()
